+++ src/models/process.ts
@@ -1,0 +1,6 @@
+export class Process {
+
+  constructor(
+    public readonly execute: () => Promise<string>,
+    public readonly terminate: () => Promise<void>) {  }
+}